Разбивает редьюсер на модули
