+++ package.json
@@ -32,12 +32,15 @@
     "node": "10"
   },
   "dependencies": {
+    "axios": "0.19.2",
     "leaflet": "1.6.0",
     "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-redux": "7.2.0",
     "react-router-dom": "5.2.0",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
