+++ src/index.js
@@ -4,7 +4,7 @@ import App from "./components/app/app.jsx";
 import offers from "./mock/offers.js";
 import cities from "./mock/cities.js";
 import {createStore} from "redux";
-import {reducer} from "./reducer.js";
+import reducer from "./reducer.js";
 import {Provider} from "react-redux";
 
 const init = () => {
@@ -17,7 +17,6 @@ const init = () => {
   ReactDOM.render(
       <Provider store={store}>
         <App
-          offers = {offers}
           cities={cities}/>
       </Provider>,
       document.querySelector(`#root`)
