+++ src/reducer.js
@@ -1,88 +1,11 @@
-import {getOffersInCity, extend, getSortedOffers} from './utils';
-import {SortType} from './const.js';
-import offers from './mock/offers';
+import {combineReducers} from "redux";
+import {reducer as data} from "./reducer/data/data.js";
+import {reducer as ui} from "./reducer/ui/ui.js";
+// import {reducer as user} from "./user/user.js";
+import NameSpace from "./reducer/name-space.js";
 
-const initialState = {
-  city: {
-    cityName: `Amsterdam`,
-    cityCoords: [52.38333, 4.9],
-  },
-  allOffers: offers,
-  currentOffers: getOffersInCity(`Amsterdam`, offers),
-  currentSortType: `Popular`,
-  activeOfferId: null,
-  activeOffer: null,
-  showSortMenu: false,
-};
-
-const ActionType = {
-  CHANGE_LOCATION: `CHANGE_LOCATION`,
-  CHANGE_CURRENT_OFFERS: `CHANGE_CURRENT_OFFERS`,
-  SORT_OFFERS: `SORT_OFFERS`,
-  SET_ACTIVE_OFFER_ID: `SET_ACTIVE_OFFER_ID`,
-  SET_ACTIVE_OFFER: `SET_ACTIVE_OFFER`,
-  SHOW_SORT_MENU: `SHOW_SORT_MENU`,
-};
-
-const ActionCreator = {
-  changeLocation: (city) => ({
-    type: ActionType.CHANGE_LOCATION,
-    payload: city,
-  }),
-  changeCurrentOffers: (cityName) => ({
-    type: ActionType.CHANGE_CURRENT_OFFERS,
-    payload: cityName,
-  }),
-  sortOffers: (sortType) => ({
-    type: ActionType.SORT_OFFERS,
-    payload: sortType,
-  }),
-  setActiveOfferId: (offerId) => ({
-    type: ActionType.SET_ACTIVE_OFFER_ID,
-    payload: offerId,
-  }),
-  setActiveOffer: (offer) => ({
-    type: ActionType.SET_ACTIVE_OFFER,
-    payload: offer,
-  }),
-  showSortMenu: (isShown) => ({
-    type: ActionType.SHOW_SORT_MENU,
-    payload: isShown,
-  })
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.CHANGE_LOCATION:
-      return extend(state, {city: action.payload});
-    case ActionType.CHANGE_CURRENT_OFFERS:
-      return extend(state, {
-        currentOffers: getOffersInCity(action.payload, state.allOffers),
-      });
-    case ActionType.SORT_OFFERS:
-      return extend(state, {
-        currentSortType: action.payload,
-        currentOffers: action.payload === SortType.POPULAR
-          ? state.currentOffers
-          : getSortedOffers(state.currentOffers, action.payload),
-      });
-    case ActionType.SET_ACTIVE_OFFER_ID:
-      return extend(state, {
-        activeOfferId: action.payload,
-      });
-
-    case ActionType.SET_ACTIVE_OFFER:
-      return extend(state, {
-        activeOffer: action.payload,
-      });
-
-    case ActionType.SHOW_SORT_MENU:
-      return extend(state, {
-        showSortMenu: action.payload,
-      });
-    default:
-      return state;
-  }
-};
-
-export {reducer, ActionType, ActionCreator};
+export default combineReducers({
+  [NameSpace.DATA]: data,
+  [NameSpace.UI]: ui,
+  // [NameSpace.USER]: user,
+});
