+++ src/reducer/user/user.js
@@ -1,0 +1,32 @@
+import {extend} from './utils';
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.NO_AUTH,
+};
+
+const ActionType = {
+  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
+};
+
+const ActionCreator = {
+  requireAuthorization: (status) => {
+    return {
+      type: ActionType.REQUIRED_AUTHORIZATION,
+      payload: status,
+    };
+  },
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.REQUIRED_AUTHORIZATION:
+      return extend(state, {
+        authorizationStatus: action.payload,
+      });
+  }
+
+  return state;
+};
+
+
+export {ActionCreator, ActionType, reducer};
